Chore: Fix Bugs

+++ src/database/migrations/1589307344336-CreateTransactions.ts
@@ -20,7 +20,7 @@ export default class CreateTransactions1589307344336
           },
           {
             name: 'value',
-            type: 'decimal',
+            type: 'integer',
           },
           {
             name: 'type',
